+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_load.cs
@@ -66,7 +66,7 @@ public static partial class SessionBaseExtensionMethods
     /// <param name="runScriptAfterLoading">    (Optional) [false] True to run script after loading. </param>
     /// <param name="deleteExisting">           (Optional) [false] True to delete the existing. </param>
     /// <param name="ignoreExisting">           (Optional) [false] True to ignore existing. </param>
-    public static void LoadScript( this Pith.SessionBase? session, string scriptName, string source, TimeSpan lineDelay,
+    public static void LoadScript( this Pith.SessionBase session, string scriptName, string source, TimeSpan lineDelay,
         bool runScriptAfterLoading = false, bool deleteExisting = false, bool ignoreExisting = false )
     {
         if ( session is null ) throw new ArgumentNullException( nameof( session ) );
@@ -91,7 +91,7 @@ public static partial class SessionBaseExtensionMethods
     /// <param name="runScriptAfterLoading">    (Optional) [false] True to run script after loading. </param>
     /// <param name="deleteExisting">           (Optional) [false] True to delete the existing. </param>
     /// <param name="ignoreExisting">           (Optional) [false] True to ignore existing. </param>
-    public static void LoadScript( this Pith.SessionBase? session, string scriptName, string[] scriptLines, TimeSpan lineDelay,
+    public static void LoadScript( this Pith.SessionBase session, string scriptName, string[] scriptLines, TimeSpan lineDelay,
         bool runScriptAfterLoading = false, bool deleteExisting = false, bool ignoreExisting = false )
     {
         if ( session is null ) throw new ArgumentNullException( nameof( session ) );