+++ changelog.md
@@ -7,6 +7,7 @@ Notable changes to the `cc.isr.vi` solution are documented in this file using th
 ## [8.1.9248] - 2025-04-27
 - Device Tsp Scrip Namespace
   - Add query firmware version methods.
+  - Remove nullable from session base argument.
 
 ## [8.1.9247] - 2025-04-26
 - Device Tsp Script Namespace